requirements.txt
pandas==1.5.0
python-decouple==3.6
requests==2.27.1
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
psycopg2==2.9.4